Extract `Header` component
