+++ app/styles/application.module.css
@@ -71,7 +71,6 @@ noscript {
     }
 }
 
-.header,
 .main,
 .footer {
     display: flex;
@@ -79,134 +78,9 @@ noscript {
     width: 100%;
 }
 
-.header-inner {
-    composes: width-limit;
-
-    display: grid;
-    grid-template-columns: auto 1fr auto;
-    align-items: center;
-    padding: 10px;
-    color: white;
-
-    a {
-        color: white; text-decoration: none;
-        &:hover { color: darken(white, 15%); }
-    }
-}
-
-.index-link {
-    display: flex;
-    align-items: center;
-
-    h1 {
-        font-size: 24px;
-    }
-}
-
-.logo {
-    width: 60px;
-    height: auto;
-    margin-right: 10px;
-}
-
-.desktop-search-form {
-    display: flex;
-    flex-grow: 1;
-
-    @media only screen and (max-width: 820px) {
-        display: none;
-    }
-}
-
-input.search {
-    font-size: 90%;
-    border: none;
-    color: black;
-    width: 100%;
-    margin-left: 16px;
-    margin-right: 16px;
-    padding: 5px 5px 5px 25px;
-    background-color: white;
-    background-image: url('/assets/search.png');
-    background-repeat: no-repeat;
-    background-position: 6px 6px;
-    background-size: 14px 15px;
-    border-radius: 15px;
-    box-shadow: none;
-}
-
-.mobile-search {
-    display: none;
-    grid-row: 2;
-    grid-column: span 3;
-    margin: 0 10px 10px;
-
-    input.search {
-        margin: 0;
-    }
-
-    @media only screen and (max-width: 820px) {
-        display: block;
-    }
-}
-
-.search-label {
-    position: absolute;
-    left: -10000px;
-    top: auto;
-    width: 1px;
-    height: 1px;
-    overflow: hidden;
-}
-
 .sep {
     margin: 0 10px;
     color: var(--separator-color);
-}
-
-.doc-links {
-    left: auto;
-    min-width: 150px;
-}
-
-.nav {
-    text-align: right;
-    margin-right: 5px;
-
-    @media only screen and (max-width: 900px) {
-        display: none;
-    }
-}
-
-.menu {
-    text-align: right;
-    margin-right: 5px;
-    flex-grow: 2;
-    display: none;
-
-    @media only screen and (max-width: 900px) {
-        display: block;
-    }
-}
-
-.menu-item-with-separator {
-    border-top: 1px solid var(--gray-border);
-}
-
-.current-user-links {
-    left: auto;
-    right: 0;
-    min-width: 200px;
-}
-
-.dropdown-button {
-    background: none;
-    border: 0;
-    padding: 10px 0;
-
-    img {
-        vertical-align: top;
-    }
 }
 
 .main {
